requirements.txt
azure-core==1.30.0
azure-data-tables==12.5.0
azure-identity==1.15.0